Alter employer avatar updating behaviour
In the db, there would be only and only avatar record for each employer record
so when it comes to updating an avatar, a checking procedure must initialized
which looks up to check whether avatar existing.

If there is an avatar it must updated, otherwise brand new avatar must be created.